Feature/101 auto stop conference

+++ src/scheduler/scheduler.service.ts
@@ -89,7 +89,26 @@ export class SchedulerService {
     );
   }
 
-  @OnEvent(ApplicationEvent.MEETING_END)
+  @OnEvent(ApplicationEvent.MEETING_CREATE)
+  /**
+   * Schedules the end of a bubble for a meeting.
+   * @param meeting - The meeting with attendees.
+   * @returns A Promise that resolves when the bubble end is scheduled.
+   */
+  public async scheduleBubbleEnd(meeting: MeetingWithAttendees): Promise<void> {
+    const bubbleEndDate = new Date(meeting.end_date);
+    const endJob = new CronJob(bubbleEndDate, async () => {
+      this.eventEmitter.emit(ApplicationEvent.MEETING_END, meeting);
+    });
+    endJob.start();
+
+    this.schedulerRegistry.addCronJob(
+      `meeting-${meeting.id}-bubble-end`,
+      endJob,
+    );
+  }
+
+  @OnEvent(ApplicationEvent.CONFERENCE_STOPPED)
   /**
    * Schedules bubble cleaning for a given bubble.
    * @param bubble - The bubble to schedule cleaning for.
@@ -114,14 +133,19 @@ export class SchedulerService {
     );
   }
 
-  @OnEvent(ApplicationEvent.MEETING_CANCEL_END)
+  @OnEvent(ApplicationEvent.CONFERENCE_STARTED)
   /**
    * Cancels the bubble cleaning for a given bubble.
    * @param bubble - The bubble for which the cleaning is to be canceled.
    * @returns A promise that resolves when the cleaning is canceled.
    */
   public async cancelBubbleCleaning(bubble: Bubble): Promise<void> {
-    this.schedulerRegistry.deleteCronJob(`bubble-${bubble.id}-cleaning`);
+    if (
+      this.schedulerRegistry.doesExist('cron', `bubble-${bubble.id}-cleaning`)
+    ) {
+      this.logger.debug(`Canceling bubble cleaning for id ${bubble.id}`);
+      this.schedulerRegistry.deleteCronJob(`bubble-${bubble.id}-cleaning`);
+    }
   }
 
   /**
@@ -134,6 +158,7 @@ export class SchedulerService {
     meetings.forEach(async (meeting) => {
       this.scheduleBubbleCreation(meeting);
       this.scheduleBubbleStart(meeting);
+      this.scheduleBubbleEnd(meeting);
     });
   }
 }
